+++ CHANGES.md
@@ -1,6 +1,7 @@
 Changes coming in 0.3.5
 
 * HSQLDB and SQLite in-memory strings are now accepted [JDBC-94](https://dev.clojure.org/jira/browse/JDBC-94).
+* Add support for readonly transactions via :read-only? [JDBC-93](https://dev.clojure.org/jira/browse/JDBC-93).
 
 Changes in 0.3.4